refactor(llama): split vocabulary out
